requirements.txt
phidata==2.7.10
google-generativeai==0.8.4
python-dotenv==1.0.1
streamlit==1.41.1
